+++ docs/assets.md
@@ -1,0 +1,7 @@
+## download_assets
+
+:::supervision.assets.downloader.download_assets
+
+## VideoAssets
+
+:::supervision.assets.list.VideoAssets
