+++ supervision/assets/__init__.py
@@ -1,2 +1,2 @@
-from .asset_list import VideoAssets
+from .list import VideoAssets
 from .downloader import download_assets
